More edits to the PR: https://github.com/CS2Surf/Timer/pull/8 (#4)
* Tidy up player stats classes

* Move Checkpoints load method to PlayerStats

* Use info_teleport_destinations for player spawn locations

* Look for info_teleport_destinations with guideline names

* Fallback if associated info_teleport_destination cannot be found

* Append `spawn_` to info_teleport_destination standards

* Resolve warnings
